Separate serach and suggestions, checks for closed sessions
