+++ android/src/main/java/com/unact/yandexmapkit/YandexSearchHandlerImpl.java
@@ -2,21 +2,15 @@ package com.unact.yandexmapkit;
 
 import android.content.Context;
 
-import androidx.annotation.NonNull;
-
 import com.yandex.mapkit.geometry.Geometry;
 import com.yandex.mapkit.geometry.Point;
 import com.yandex.mapkit.geometry.BoundingBox;
 import com.yandex.mapkit.search.SearchOptions;
 import com.yandex.mapkit.search.Session;
 import com.yandex.mapkit.search.Snippet;
-import com.yandex.mapkit.search.SuggestItem;
 import com.yandex.mapkit.search.SearchFactory;
 import com.yandex.mapkit.search.SearchManagerType;
-import com.yandex.mapkit.search.SuggestOptions;
 import com.yandex.mapkit.search.SearchManager;
-import com.yandex.mapkit.search.SuggestSession;
-import com.yandex.runtime.Error;
 
 import java.util.ArrayList;
 import java.util.HashMap;
@@ -31,18 +25,15 @@ import io.flutter.plugin.common.MethodChannel.Result;
 
 public class YandexSearchHandlerImpl implements MethodCallHandler {
 
-  private MethodChannel methodChannel;
   private final SearchManager searchManager;
   private final BinaryMessenger binaryMessenger;
 
-  private Map<Integer, SuggestSession>  		suggestSessionsById = new HashMap<>();
   private Map<Integer, YandexSearchSession> searchSessionsById  = new HashMap<>();;
 
-  public YandexSearchHandlerImpl(Context context, MethodChannel channel, BinaryMessenger messenger) {
+  public YandexSearchHandlerImpl(Context context, BinaryMessenger messenger) {
 
     SearchFactory.initialize(context);
 
-    methodChannel 	= channel;
     searchManager 	= SearchFactory.getInstance().createSearchManager(SearchManagerType.COMBINED);
 		binaryMessenger	= messenger;
   }
@@ -50,14 +41,6 @@ public class YandexSearchHandlerImpl implements MethodCallHandler {
   @Override
   public void onMethodCall(MethodCall call, Result result) {
     switch (call.method) {
-      case "getSuggestions":
-        getSuggestions(call);
-        result.success(null);
-        break;
-      case "cancelSuggestSession":
-        cancelSuggestSession(call);
-        result.success(null);
-        break;
       case "searchByText":
         searchByText(call, result);
         break;
@@ -68,33 +51,6 @@ public class YandexSearchHandlerImpl implements MethodCallHandler {
         result.notImplemented();
         break;
     }
-  }
-
-  @SuppressWarnings("unchecked")
-  private void cancelSuggestSession(MethodCall call) {
-    Map<String, Object> params = ((Map<String, Object>) call.arguments);
-    final int listenerId = ((Number) params.get("listenerId")).intValue();
-
-    suggestSessionsById.remove(listenerId);
-  }
-
-  @SuppressWarnings("unchecked")
-  private void getSuggestions(MethodCall call) {
-    Map<String, Object> params = ((Map<String, Object>) call.arguments);
-    final int listenerId = ((Number) params.get("listenerId")).intValue();
-    String formattedAddress = (String) params.get("formattedAddress");
-    BoundingBox boundingBox = new BoundingBox(
-      new Point(((Double) params.get("southWestLatitude")), ((Double) params.get("southWestLongitude"))),
-      new Point(((Double) params.get("northEastLatitude")), ((Double) params.get("northEastLongitude")))
-    );
-    Boolean suggestWords = ((Boolean) params.get("suggestWords"));
-    SuggestSession suggestSession = searchManager.createSuggestSession();
-    SuggestOptions suggestOptions = new SuggestOptions();
-    suggestOptions.setSuggestTypes(((Number) params.get("suggestType")).intValue());
-    suggestOptions.setSuggestWords(suggestWords);
-
-    suggestSession.suggest(formattedAddress, boundingBox, suggestOptions, new YandexSuggestListener(listenerId));
-    suggestSessionsById.put(listenerId, suggestSession);
   }
 
   public void searchByText(MethodCall call, MethodChannel.Result result) {
@@ -232,47 +188,4 @@ public class YandexSearchHandlerImpl implements MethodCallHandler {
 			searchSessionsById.remove(sessionId);
 		}
 	}
-
-  private class YandexSuggestListener implements SuggestSession.SuggestListener {
-
-    public YandexSuggestListener(int id) {
-      listenerId = id;
-    }
-
-    private int listenerId;
-
-    @Override
-    public void onResponse(@NonNull List<SuggestItem> suggestItems) {
-      List<Map<String, Object>> suggests = new ArrayList<>();
-
-      for (SuggestItem suggestItemResult : suggestItems) {
-        Map<String, Object> suggestMap = new HashMap<>();
-        suggestMap.put("title", suggestItemResult.getTitle().getText());
-        if(suggestItemResult.getSubtitle() != null) {
-          suggestMap.put("subtitle", suggestItemResult.getSubtitle().getText());
-        }
-        if(suggestItemResult.getDisplayText() != null) {
-          suggestMap.put("displayText", suggestItemResult.getDisplayText());
-        }
-        suggestMap.put("searchText", suggestItemResult.getSearchText());
-        suggestMap.put("type", suggestItemResult.getType().ordinal());
-        suggestMap.put("tags", suggestItemResult.getTags());
-
-        suggests.add(suggestMap);
-      }
-
-      Map<String, Object> arguments = new HashMap<>();
-      arguments.put("listenerId", listenerId);
-      arguments.put("response", suggests);
-
-      methodChannel.invokeMethod("onSuggestListenerResponse", arguments);
-    }
-
-    @Override
-    public void onError(@NonNull Error error) {
-			Map<String, Object> arguments = new HashMap<>();
-			arguments.put("listenerId", listenerId);
-			methodChannel.invokeMethod("onSuggestListenerError", arguments);
-    }
-  }
 }
